fix(consensus): BIP-68 version gate must compare unsigned (Core uint32_t parity)
Core stores tx.version as uint32_t and computes fEnforceBIP68 = version >= 2
UNSIGNED (consensus/tx_verify.cpp:51), so a high-bit version (e.g. 0x80000002)
still enforces BIP-68. blockbrew stores MsgTx.Version as int32 and the connect-block
sequence-lock gate (CalculateSequenceLocks, txvalidation.go) compared signed
(tx.Version less than 2), treating 0x80000002 as negative, so BIP-68 enforcement was
SKIPPED and the node FALSE-ACCEPTS a tx whose relative timelock is unmet, which Core
rejects (a chain split). The OP_CSV path was already fixed unsigned
(script/opcodes_impl.go, with a comment naming this exact pitfall) but this gate was
missed -- an asymmetric fix.

Compare uint32(tx.Version) against 2. Found by the differential bug-hunt (9/10 impls
affected); Core-verified. Verified pre/post: the new TestCalculateSequenceLocks case
asserts a high-bit version computes the height lock (109) -- it FAILS on the signed
compare (MinHeight = -1) and PASSES on the fix.

+++ internal/consensus/txvalidation.go
@@ -500,8 +500,13 @@ func CalculateSequenceLocks(tx *wire.MsgTx, prevHeights []int32, getMTP MTPLooku
 		MinTime:   -1,
 	}
 
-	// BIP68 only applies to version >= 2 transactions.
-	if tx.Version < 2 {
+	// BIP68 only applies to version >= 2 transactions. Core stores version as
+	// uint32_t and compares unsigned (consensus/tx_verify.cpp:51), so a high-bit
+	// version (e.g. 0x80000002) still enforces BIP68. tx.Version is int32 here, so
+	// a signed `< 2` would treat 0x80000002 as negative and SKIP enforcement,
+	// false-accepting a tx with an unmet relative timelock (chain split). Compare
+	// unsigned to match Core -- same as the OP_CSV path (script/opcodes_impl.go).
+	if uint32(tx.Version) < 2 {
 		return lock
 	}
